Place more filters on records list view
This view sends data from public-facing input to the database, and the order parameter in particular is vulnerable to SQL injection. It seems that within the context of the query, order needs to match one of the available column names, so I've made it so that those are the only parameters that are accepted. Any order parameter that isn't one of those is ignored, which prevents any malicious code from being executed.

Out of an abundance of caution, I've also quoted parameters that expect int values to a quote function for ints, and everything else besides the extent parameter gets quoted as a string. Extent gets map geometry data, and was vulnerable to sql injection prior to Neatline 2.6.3, but that is now validated at the point where the geometry is used.